Refactored repository part 1.

+++ db/migration/V0_0_0__create_schema.sql
@@ -5,7 +5,7 @@ CREATE SCHEMA IF NOT EXISTS fs_auto;
 CREATE SEQUENCE fs_auto.person_seq START WITH 1 INCREMENT BY 1;
 
 CREATE TABLE fs_auto.person
-(id NUMERIC NOT NULL
+(id BIGINT NOT NULL
 ,name CHARACTER VARYING(255)
 ,cpf CHARACTER VARYING(20)
 ,cell_phone CHARACTER VARYING(20)
@@ -22,7 +22,7 @@ ALTER TABLE fs_auto.person ALTER COLUMN id SET DEFAULT NEXTVAL('fs_auto.person_s
 CREATE SEQUENCE fs_auto.course_seq START WITH 1 INCREMENT BY 1;
 
 CREATE TABLE fs_auto.course
-(id NUMERIC NOT NULL
+(id BIGINT NOT NULL
 ,name CHARACTER VARYING(255)
 ,description CHARACTER VARYING(255)
 ,registration_date TIMESTAMP
@@ -35,8 +35,8 @@ ALTER TABLE fs_auto.course ALTER COLUMN id SET DEFAULT NEXTVAL('fs_auto.course_s
 CREATE SEQUENCE fs_auto.school_class_seq START WITH 1 INCREMENT BY 1;
 
 CREATE TABLE fs_auto.school_class
-(id NUMERIC NOT NULL
-,course_id NUMERIC NOT NULL
+(id BIGINT NOT NULL
+,course_id BIGINT NOT NULL
 ,start_date TIMESTAMP
 ,end_date TIMESTAMP
 ,registration_date TIMESTAMP
@@ -50,8 +50,8 @@ ALTER TABLE fs_auto.school_class ALTER COLUMN id SET DEFAULT NEXTVAL('fs_auto.sc
 
 ---------- STUDENT ----------
 CREATE TABLE fs_auto.student
-(id NUMERIC NOT NULL
-,school_class_id NUMERIC
+(id BIGINT NOT NULL
+,school_class_id BIGINT
 ,CONSTRAINT pk_student PRIMARY KEY (id)
 ,CONSTRAINT fk_student_person FOREIGN KEY (id) REFERENCES fs_auto.person (id)
 ,CONSTRAINT fk_student_school_class FOREIGN KEY (school_class_id) REFERENCES fs_auto.school_class (id));
@@ -61,8 +61,8 @@ CREATE INDEX idx_student_school_class ON fs_auto.student(school_class_id);
 
 ---------- TEACHER ----------
 CREATE TABLE fs_auto.teacher
-(id NUMERIC NOT NULL
-,course_id NUMERIC
+(id BIGINT NOT NULL
+,course_id BIGINT
 ,CONSTRAINT pk_teacher PRIMARY KEY (id)
 ,CONSTRAINT fk_teacher_person FOREIGN KEY (id) REFERENCES fs_auto.person (id)
 ,CONSTRAINT fk_teacher_course FOREIGN KEY (course_id) REFERENCES fs_auto.course (id));
@@ -74,7 +74,7 @@ CREATE INDEX idx_teacher_course ON fs_auto.teacher(course_id);
 CREATE SEQUENCE fs_auto.profile_seq START WITH 1 INCREMENT BY 1;
 
 CREATE TABLE fs_auto.profile
-(id NUMERIC NOT NULL
+(id BIGINT NOT NULL
 ,name CHARACTER VARYING(255)
 ,code CHARACTER VARYING(5)
 ,access_permission CHARACTER VARYING(50)
@@ -87,11 +87,11 @@ ALTER TABLE fs_auto.profile ALTER COLUMN id SET DEFAULT NEXTVAL('fs_auto.profile
 CREATE SEQUENCE fs_auto.app_user_seq START WITH 1 INCREMENT BY 1;
 
 CREATE TABLE fs_auto.app_user
-(id NUMERIC NOT NULL
+(id BIGINT NOT NULL
 ,name CHARACTER VARYING(255)
 ,password CHARACTER VARYING(255)
-,profile_id NUMERIC
-,person_id NUMERIC
+,profile_id BIGINT
+,person_id BIGINT
 ,start_date TIMESTAMP
 ,end_date TIMESTAMP
 ,registration_date TIMESTAMP
